pyproject.toml
[tool.poetry]
name = "apache_grpc"
version = "0.1.0"
description = ""
authors = ["Eugene Morozov <jmv@emorozov.net>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.3"
grpcio = "^1.50.0"

[tool.poetry.group.dev.dependencies]
grpcio-tools = "^1.50.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
